+++ app/routes/projects.py
@@ -1,6 +1,7 @@
 from flask import make_response
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from app.controllers.auth import is_admin
 from app.controllers.projects import (
     load_projects,
     add_project,
@@ -31,6 +32,10 @@ class Projects(Resource):
 
     @jwt_required
     def post(self):
+        identity = get_jwt_identity()
+        if not is_admin(identity):
+            return make_response({'success': False, 'message': 'unauthorized'}, 401)
+
         project_data = self.reqparse.parse_args()
         new_project = add_project(project_data)
 
@@ -41,6 +46,10 @@ class Projects(Resource):
 
     @jwt_required
     def delete(self):
+        identity = get_jwt_identity()
+        if not is_admin(identity):
+            return make_response({'success': False, 'message': 'unauthorized'}, 401)
+
         total = delete_projects()
         return make_response({'success': True, 'total_deleted': total}, 200)
 
@@ -64,6 +73,10 @@ class Project(Resource):
 
     @jwt_required
     def put(self, id):
+        identity = get_jwt_identity()
+        if not is_admin(identity):
+            return make_response({'success': False, 'message': 'unauthorized'}, 401)
+
         project = get_project(id)
         if not project:
             return make_response({'success': False, 'message': 'project not found'}, 404)
@@ -77,6 +90,10 @@ class Project(Resource):
 
     @jwt_required
     def delete(self, id):
+        identity = get_jwt_identity()
+        if not is_admin(identity):
+            return make_response({'success': False, 'message': 'unauthorized'}, 401)
+
         project = get_project(id)
         if not project:
             return make_response({'success': False, 'message': 'project not found'}, 404)
