+++ app/routes/skills.py
@@ -1,6 +1,7 @@
 from flask import make_response
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from app.controllers.auth import is_admin
 from app.controllers.skills import (
     add_skill,
     fetch_skills,
@@ -19,8 +20,20 @@ class Skills(Resource):
         self.reqparse.add_argument(
             'stack', type=list, location='json', required=True)
 
+    def get(self):
+        skills = fetch_skills()
+
+        if not skills:
+            return make_response({'success': False}, 400)
+
+        return make_response({'success': True, 'skills': skills}, 200)
+
     @jwt_required
     def post(self):
+        identity = get_jwt_identity()
+        if not is_admin(identity):
+            return make_response({'success': False, 'message': 'unauthorized'}, 401)
+
         args = self.reqparse.parse_args()
         title = args.get('title')
         stack = args.get('stack')
@@ -30,14 +43,6 @@ class Skills(Resource):
             return make_response({'success': False}, 400)
 
         return make_response({'success': True, 'skill': skill}, 200)
-
-    def get(self):
-        skills = fetch_skills()
-
-        if not skills:
-            return make_response({'success': False}, 400)
-
-        return make_response({'success': True, 'skills': skills}, 200)
 
 
 class Skill(Resource):
@@ -59,6 +64,10 @@ class Skill(Resource):
 
     @jwt_required
     def put(self, id):
+        identity = get_jwt_identity()
+        if not is_admin(identity):
+            return make_response({'success': False, 'message': 'unauthorized'}, 401)
+
         args = self.reqparse.parse_args()
         stack = args.get('stack')
         full = args.get('full')
@@ -71,6 +80,10 @@ class Skill(Resource):
 
     @jwt_required
     def delete(self, id):
+        identity = get_jwt_identity()
+        if not is_admin(identity):
+            return make_response({'success': False, 'message': 'unauthorized'}, 401)
+
         if not delete_skill_set(id):
             return make_response({'success': False, 'message': 'not found'}, 404)
